feat(#2129): extract governance brick from services/ to bricks/
Move governance domain (~2,813 LOC, 14 files) to self-contained brick
at nexus/bricks/governance/ per Nexus Lego Architecture (§3).

Infrastructure:
- Move ORM base to nexus.contracts.db_base (re-export shim at old path)
- Move TransactionProtocol to nexus.contracts.types
- Create nexus.utils.async_helpers for fire_and_forget

Brick structure (17 files):
- Models, DB models, protocols, approval/ subpackage
- 6 services: anomaly, collusion, graph, response, wrapper, trust_math
- Extracted json_utils.py (DRY: 6 occurrences) and converters.py

Code quality:
- GovernanceProtocol: replaced Any with concrete types
- ApprovalStatus.NONE enum replaces string literals
- GovernanceSnapshot dataclass avoids double edge loading
- Cache eviction fix with periodic sweep every 100 lookups
- scipy sparse matrices for EigenTrust (prevents dense NxN OOM)

Tests: 51 new tests across 4 test files (math, cache, approval)
Lego compliance: zero runtime imports from nexus.core or other bricks
